# Slight performance improvement for :class:`.ArrowVectorField` and Bezier curve computation