--- a/manim/utils/bezier.py
+++ b/manim/utils/bezier.py
@@ -46,6 +46,23 @@ def bezier(
         function describing the bezier curve.
     """
     n = len(points) - 1
+
+    # Cubic Bezier curve
+    if n == 3:
+        return (
+            lambda t: (1 - t) ** 3 * points[0]
+            + 3 * t * (1 - t) ** 2 * points[1]
+            + 3 * (1 - t) * t**2 * points[2]
+            + t**3 * points[3]
+        )
+    # Quadratic Bezier curve
+    if n == 2:
+        return (
+            lambda t: (1 - t) ** 2 * points[0]
+            + 2 * t * (1 - t) * points[1]
+            + t**2 * points[2]
+        )
+
     return lambda t: sum(
         ((1 - t) ** (n - k)) * (t**k) * choose(n, k) * point
         for k, point in enumerate(points)