<!-- Thank you for contributing to Manim! Learn more about the process in our contributing guidelines: https://docs.manim.community/en/latest/contributing.html -->

## Overview: What does this pull request change?
Minor tweaks to improve the performance of creating a ArrowVectorField.

Also tested but useless:
- Changing to an explicit definition of bezier which does not return a lambda but takes the points and an alpha
- making bezier vectorized to work with stacks of points
- reducing calls to bezier

<details><summary>Changes over time which turned out to be useless sometimes</summary>

```diff

diff --git a/manim/mobject/geometry/line.py b/manim/mobject/geometry/line.py
index cb8d7188..4fe9bb7e 100644
--- a/manim/mobject/geometry/line.py
+++ b/manim/mobject/geometry/line.py
@@ -102,12 +102,12 @@ class Line(TipableVMobject):
             The direction.
         """
         if isinstance(mob_or_point, (Mobject, OpenGLMobject)):
-            mob = mob_or_point
+            # mob_or_point is now a mobject
             if direction is None:
-                return mob.get_center()
+                return mob_or_point.get_center()
             else:
-                return mob.get_boundary_point(direction)
-        return np.array(mob_or_point)
+                return mob_or_point.get_boundary_point(direction)
+        return np.asarray(mob_or_point)
 
     def set_path_arc(self, new_value):
         self.path_arc = new_value
diff --git a/manim/mobject/types/vectorized_mobject.py b/manim/mobject/types/vectorized_mobject.py
index d7bb7313..6ccb676a 100644
--- a/manim/mobject/types/vectorized_mobject.py
+++ b/manim/mobject/types/vectorized_mobject.py
@@ -14,6 +14,7 @@ __all__ = [
 import itertools as it
 import sys
 import typing
+from functools import cache
 from typing import Optional, Sequence, Union
 
 import colour
@@ -31,6 +32,8 @@ from ...constants import *
 from ...mobject.mobject import Mobject
 from ...utils.bezier import (
     bezier,
+    explicit_bezier,
+    explicit_bezier_vectorized,
     get_smooth_handle_points,
     integer_interpolate,
     interpolate,
@@ -1062,6 +1065,7 @@ class VMobject(Mobject):
         nppcc = self.n_points_per_cubic_curve
         return self.points[nppcc * n : nppcc * (n + 1)]
 
+    @deprecated
     def get_nth_curve_function(self, n: int) -> typing.Callable[[float], np.ndarray]:
         """Returns the expression of the nth curve.
 
@@ -1099,17 +1103,23 @@ class VMobject(Mobject):
         if sample_points is None:
             sample_points = 10
 
-        curve = self.get_nth_curve_function(n)
-        points = np.array([curve(a) for a in np.linspace(0, 1, sample_points)])
+        nth_points = self.get_nth_curve_points(n)
+        points = np.array(
+            [bezier(nth_points)(a) for a in np.linspace(0, 1, sample_points)]
+        )
+        # print(f"Old: {points}")
+        # points = explicit_bezier_vectorized(
+        #     nth_points, np.linspace(0, 1, sample_points)
+        # )
+        # print(f"New: {new_points}")
+        # print(points.shape, new_points.shape)
         diffs = points[1:] - points[:-1]
-        norms = np.apply_along_axis(np.linalg.norm, 1, diffs)
+        norms = np.linalg.norm(diffs, axis=1)
 
         return norms
 
     def get_nth_curve_length(
-        self,
-        n: int,
-        sample_points: Optional[int] = None,
+        self, n: int, sample_points: Optional[int] = None
     ) -> float:
         """Returns the (approximate) length of the nth curve.
 
@@ -1125,11 +1135,25 @@ class VMobject(Mobject):
         length : :class:`float`
             The length of the nth curve.
         """
+        return np.sum(self.get_nth_curve_length_pieces(n, sample_points=sample_points))
 
-        _, length = self.get_nth_curve_function_with_length(n, sample_points)
+    def get_curve_lengths(self, sample_points: Optional[int] = None) -> np.ndarray:
+        """Returns the array of lengths of the curves
 
-        return length
+        Parameters
+        ----------
+        sample_points
+            The number of points to sample to find the length.
+
+        Returns
+        -------
+        list_length : :class: `np.ndarray`
+            list of lengths of the curves.
+        """
+        for i in range(self.get_num_curves()):
+            yield self.get_nth_curve_length(i, sample_points=sample_points)
 
+    @deprecated
     def get_nth_curve_function_with_length(
         self,
         n: int,
@@ -1169,6 +1193,7 @@ class VMobject(Mobject):
         nppcc = self.n_points_per_cubic_curve
         return len(self.points) // nppcc
 
+    @deprecated
     def get_curve_functions(
         self,
     ) -> typing.Iterable[typing.Callable[[float], np.ndarray]]:
@@ -1185,6 +1210,7 @@ class VMobject(Mobject):
         for n in range(num_curves):
             yield self.get_nth_curve_function(n)
 
+    @deprecated
     def get_curve_functions_with_lengths(
         self, **kwargs
     ) -> typing.Iterable[typing.Tuple[typing.Callable[[float], np.ndarray], float]]:
@@ -1234,19 +1260,19 @@ class VMobject(Mobject):
         if alpha == 1:
             return self.points[-1]
 
-        curves_and_lengths = tuple(self.get_curve_functions_with_lengths())
+        lengths = tuple(self.get_curve_lengths())
 
-        target_length = alpha * sum(length for _, length in curves_and_lengths)
+        target_length = alpha * sum(lengths)
         current_length = 0
 
-        for curve, length in curves_and_lengths:
+        for i, length in enumerate(lengths):
             if current_length + length >= target_length:
                 if length != 0:
                     residue = (target_length - current_length) / length
                 else:
                     residue = 0
 
-                return curve(residue)
+                return explicit_bezier(self.get_nth_curve_points(i), residue)
 
             current_length += length
 
@@ -1375,7 +1401,7 @@ class VMobject(Mobject):
 
         return sum(
             length
-            for _, length in self.get_curve_functions_with_lengths(
+            for length in self.get_curve_lengths(
                 sample_points=sample_points_per_curve,
             )
         )
diff --git a/manim/mobject/vector_field.py b/manim/mobject/vector_field.py
index 2c09edcb..89786704 100644
--- a/manim/mobject/vector_field.py
+++ b/manim/mobject/vector_field.py
@@ -575,10 +575,27 @@ class ArrowVectorField(VectorField):
         x_range = np.arange(*self.x_range)
         y_range = np.arange(*self.y_range)
         z_range = np.arange(*self.z_range)
-        for x, y, z in it.product(x_range, y_range, z_range):
-            self.add(self.get_vector(x * RIGHT + y * UP + z * OUT))
+
+        # vecs = [self.get_vector(x * RIGHT + y * UP + z * OUT) for x, y, z in it.product(x_range, y_range, z_range)]
+        vecs = self.get_vectors(np.stack(list(it.product(x_range, y_range, z_range))))
+        self.add(*vecs)
         self.set_opacity(self.opacity)
 
+    def get_vectors(self, points: np.ndarray):
+        output = np.asarray([self.func(point) for point in points])
+        norm = np.linalg.norm(output, axis=1)
+        mask = norm != 0
+        newvals = (output[mask].T * self.length_func(norm[mask]) / norm[mask]).T
+        output[mask] = newvals
+        color_func = (
+            (lambda x: self.color) if self.single_color else self.pos_to_color
+        )  # TODO: Still calls self.func in pos_to_color -> there should be a vec_to_color instead
+        vectors = [
+            Vector(vec, **self.vector_config).shift(point).set_color(color_func(point))
+            for vec, point in zip(output, points)
+        ]
+        return vectors
+
     def get_vector(self, point: np.ndarray):
         """Creates a vector in the vector field.
 
@@ -594,7 +611,7 @@ class ArrowVectorField(VectorField):
             Additional arguments to be passed to the :class:`~.Vector` constructor
 
         """
-        output = np.array(self.func(point))
+        output = np.asarray(self.func(point))
         norm = np.linalg.norm(output)
         if norm != 0:
             output *= self.length_func(norm) / norm
diff --git a/manim/utils/bezier.py b/manim/utils/bezier.py
index 6f8187ee..1186adcd 100644
--- a/manim/utils/bezier.py
+++ b/manim/utils/bezier.py
@@ -2,6 +2,8 @@
 
 from __future__ import annotations
 
+import scipy
+
 __all__ = [
     "bezier",
     "partial_bezier_points",
@@ -21,15 +23,15 @@ __all__ = [
 
 
 import typing
-from functools import reduce
+from functools import cache, reduce
 
 import numpy as np
-from scipy import linalg
+from scipy import linalg, special
 
+from ..utils.deprecation import deprecated
 from ..utils.simple_functions import choose
 from ..utils.space_ops import cross2d, find_intersection
 
-
 def bezier(
     points: np.ndarray,
 ) -> typing.Callable[[float], int | typing.Iterable]:
@@ -69,6 +71,75 @@ def bezier(
     )
 
 
+def choose_vectorized(n, ks):
+    return special.factorial(n) / (special.factorial(ks) * special.factorial(n - ks))
+
+
+def explicit_bezier(
+    points: np.ndarray,
+    t: float,
+) -> np.ndarray:
+    """
+    Given a set of points, and a parameter t, return the
+    point on the curve at t.
+    """
+    n = len(points) -1
+
+    # Cubic Bezier curve
+    if n == 3:
+        return (1 - t) ** 3 * points[0] + 3 * t * (1 - t) ** 2 * points[1] + 3 * (1 - t) * t ** 2 * points[2] + t ** 3 * points[3]
+    # Quadratic Bezier curve
+    if n == 2:
+        return (1 - t) ** 2 * points[0] + 2 * t * (1 - t) * points[1] + t ** 2 * points[2]
+
+    return sum(
+        ((1 - t) ** (n - k)) * (t**k) * choose(n, k) * point
+        for k, point in enumerate(points)
+    )
+
+
+def explicit_bezier_vectorized(points, ts):
+    # ts = ts[:2]
+    n = len(points) - 1
+    p_len = points.shape[1]
+    ks = np.arange(n + 1)
+
+    reference = [
+        [
+            ((1 - t) ** (n - k)) * (t**k) * choose(n, k) * point
+            for k, point in enumerate(points)
+        ]
+        for t in ts
+    ]
+    reference = [
+        sum(
+            ((1 - t) ** (n - k)) * (t**k) * choose(n, k) * point
+            for k, point in enumerate(points)
+        )
+        for t in ts
+    ]
+    # print(f"Reference: {np.stack(reference)}")
+
+    ts_len = len(ts)
+    # print()
+    extended_ks = np.repeat([ks], ts_len, axis=0)
+    extended_ts = np.repeat([ts], n + 1, axis=0).T
+    eq = (
+        np.power(1 - extended_ts, n - ks)
+        * np.power(extended_ts, ks)
+        * choose_vectorized(n, ks)
+    )
+    # print(f"EQ1 {eq[:,:, None]}")
+    eq = np.repeat(eq[:, :, None], p_len, axis=2) * points
+    eq = np.sum(eq, axis=1)
+    # print(eq.shape, points.shape)
+    # print(np.array(reference).shape)
+    # print(f"EQ2 {eq}")
+    # print(np.array_equal(reference, eq))
+    # tmp = ((1 - ts) ** (n - ks) * (ts ** ks) * choose_vectorized(n, ks)) @ points
+    return eq
+
+
 def partial_bezier_points(points: np.ndarray, a: float, b: float) -> np.ndarray:
     """Given an array of points which define bezier curve, and two numbers 0<=a<b<=1, return an array of the same size,
     which describes the portion of the original bezier curve on the interval [a, b].
@@ -92,9 +163,11 @@ def partial_bezier_points(points: np.ndarray, a: float, b: float) -> np.ndarray:
     if a == 1:
         return [points[-1]] * len(points)
 
-    a_to_1 = np.array([bezier(points[i:])(a) for i in range(len(points))])
+    a_to_1 = np.array([explicit_bezier(points[i:], a) for i in range(len(points))])
     end_prop = (b - a) / (1.0 - a)
-    return np.array([bezier(a_to_1[: i + 1])(end_prop) for i in range(len(points))])
+    return np.array(
+        [explicit_bezier(a_to_1[: i + 1], end_prop) for i in range(len(points))]
+    )
 
 
 # Shortened version of partial_bezier_points just for quadratics,
@@ -375,8 +448,8 @@ def get_quadratic_approximation_of_cubic(a0, h0, h1, a1):
     t_mid = t_mid.repeat(n).reshape((m, n))
 
     # Compute bezier point and tangent at the chosen value of t
-    mid = bezier([a0, h0, h1, a1])(t_mid)
-    Tm = bezier([h0 - a0, h1 - h0, a1 - h1])(t_mid)
+    mid = explicit_bezier([a0, h0, h1, a1], t_mid)
+    Tm = explicit_bezier([h0 - a0, h1 - h0, a1 - h1], t_mid)
 
     # Intersection between tangent lines at end points
     # and tangent in the middle
diff --git a/manim/utils/iterables.py b/manim/utils/iterables.py
index 51a9a6ff..afcf6d99 100644
--- a/manim/utils/iterables.py
+++ b/manim/utils/iterables.py
@@ -182,10 +182,12 @@ def make_even(iterable_1: Iterable, iterable_2: Iterable) -> tuple[list, list]:
         # ([1, 1, 1, 2, 2], [3, 4, 5, 6, 7])
     """
     list_1, list_2 = list(iterable_1), list(iterable_2)
-    length = max(len(list_1), len(list_2))
+    len_list_1 = len(list_1)
+    len_list_2 = len(list_2)
+    length = max(len_list_1, len_list_2)
     return (
-        [list_1[(n * len(list_1)) // length] for n in range(length)],
-        [list_2[(n * len(list_2)) // length] for n in range(length)],
+        [list_1[(n * len_list_1) // length] for n in range(length)],
+        [list_2[(n * len_list_2) // length] for n in range(length)],
     )
 
 
diff --git a/manim/utils/rate_functions.py b/manim/utils/rate_functions.py
index 261199b5..d39d98fd 100644
--- a/manim/utils/rate_functions.py
+++ b/manim/utils/rate_functions.py
@@ -109,7 +109,7 @@ from math import sqrt
 
 import numpy as np
 
-from ..utils.bezier import bezier
+from ..utils.bezier import explicit_bezier
 from ..utils.simple_functions import sigmoid
 
 
@@ -200,7 +200,7 @@ def running_start(
     t: float,
     pull_factor: float = -0.5,
 ) -> typing.Iterable:  # what is func return type?
-    return bezier([0, 0, pull_factor, pull_factor, 1, 1, 1])(t)
+    return explicit_bezier([0, 0, pull_factor, pull_factor, 1, 1, 1], t)
 
 
 def not_quite_there(
diff --git a/manim/utils/simple_functions.py b/manim/utils/simple_functions.py
index 2a4a7afe..e957e2ca 100644
--- a/manim/utils/simple_functions.py
+++ b/manim/utils/simple_functions.py
@@ -91,7 +91,7 @@ def choose(n: int, k: int) -> int:
     - https://en.wikipedia.org/wiki/Combination
     - https://docs.scipy.org/doc/scipy/reference/generated/scipy.special.comb.html
     """
-    return special.comb(n, k, exact=True)
+    return special.comb(n, k, exact=False)
 
 
 def clip(a, min_a, max_a):

```
</details>

<!-- If there is more information than the PR title that should be added to our release changelog, add it in the following changelog section. This is optional, but recommended for larger pull requests. -->

## Changes
<!--changelog-start-->
- changed to list_comprehension for get_vector and used *vecs in add to avoid multiple calls of self.add
- changed implementation of bezier in bezier.py to also return explicit cubic and quadratic bezier functions instead of the general bezier function
- removed duplicate len calls in make_even in iterables.py
<!--changelog-end-->

## Motivation and Explanation: Why and how do your changes improve the library?
<!-- Optional for bugfixes, small enhancements, and documentation-related PRs. Otherwise, please give a short reasoning for your changes. -->
I just wanted to create a vector field and it was horribly slow... and it still is but now it's slightly faster than before

## Further Information and Comments
<!-- If applicable, put further comments for the reviewers here. -->

```py
from manim import *
    with cProfile.Profile() as pr:
        def step(state:np.ndarray):
            return np.array([state[1],(-9.8/10)*np.sin(state[0])-((0.1*state[1]/10)*10**2), 0])

        pr.enable()
        ArrowVectorField(func=step, x_range=[-4,2,0.1],y_range=[-1.5,1.5,0.1])
        pr.disable()
    pr.dump_stats("bench_opt.stats")
```

### Before
![grafik](https://user-images.githubusercontent.com/17803932/169649215-62ede908-fc5d-443a-9f71-97efc211b61b.png)

### After
![grafik](https://user-images.githubusercontent.com/17803932/169649288-306c66f3-3678-4bf3-b63e-73bfe294509c.png)


<!-- Thank you again for contributing! Do not modify the lines below, they are for reviewers. -->
## Reviewer Checklist
- [ ] The PR title is descriptive enough for the changelog, and the PR is labeled correctly
- [ ] If applicable: newly added non-private functions and classes have a docstring including a short summary and a PARAMETERS section
- [ ] If applicable: newly added functions and classes are tested